+++ dev/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.5](https://github.com/samchen08/vtj.pro/compare/dev-web@0.19.4...dev-web@0.19.5) (2026-08-11)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.19.4](https://github.com/samchen08/vtj.pro/compare/dev-web@0.19.3...dev-web@0.19.4) (2026-08-09)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
 ## [0.19.3](https://github.com/samchen08/vtj.pro/compare/dev-web@0.19.2...dev-web@0.19.3) (2026-08-09)
 
 **Note:** Version bump only for package dev-web